Fix BOM and file cleanup
